refactor user model/controller, fix mistakes - auth/auth-controller.js   rename functions - middleware/express.js   fix reference to config - server.js   fix reference to express middleware - user/user-controller.js   rename functions - user/user-model.js   rename functions - user/user-routes.js   rename functions
