+++ server/server.js
@@ -1,6 +1,6 @@
 var express = require('express');
-var config = require('./config/config');
-var configExpress = require('./config/express');
+var config = require('./config');
+var configExpress = require('./middleware/express');
 var routes =  require('./routes');
 
 var app      = express();
