+++ server/user/user-routes.js
@@ -6,8 +6,8 @@ var controller = require('./user-controller.js');
 //app.route('/users/:id').get(authentication.isAuthenticated, user.findUser);
 //app.route('/users').get(authentication.isAuthenticated, user.findUsers);
 
-routes.get('/', controller.getAll);
-routes.get('/:id', controller.getOneById);
+routes.get('/', controller.all);
+routes.get('/:id', controller.byID);
 routes.post('/create', controller.create);
 
 module.exports = routes;
